backend/photo-process/requirements.txt
requests==2.31.0
Pillow==10.2.0
piexif==1.1.3
